+++ src/main/java/com/study/simpleboard/dto/request/PostCreateRequest.java
@@ -1,7 +1,6 @@
-package com.study.simpleboard.dto;
+package com.study.simpleboard.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -10,7 +9,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 @EqualsAndHashCode
-public class PostCreateReq {
+public class PostCreateRequest {
     // 나중에 인증 구현 후 수정 예정이라 임시로 만듦
     @Positive
     private final Long userId;
